refactor: cleanup unnecessary files
