+++ packages/wallet-management/src/index.ts
@@ -1,5 +1,5 @@
 export * from './LiFiWalletManagement';
 export * from './walletAutomation';
 export * from './walletIcons';
-export * from './walletProviders';
+export * from './wallets';
 export * from './wallets';
